logout function, fixed bug protectedRoute toast
bug fixed:
only once authenticated, will remove the error message only if the error message is "Login first to access this resource." if not it will have a bug that removes all error from other components.

Other components are using useEffect to check the error as well as a dependency [] on the useEffect. It has to be implemented this way using useEffect because only when you dispatch an action, the error state will be updated. you cannot just check for error before you dispatch.